requirements.txt
# Core LLM and Embedding
langchain-mistralai
langchain-core
langchain-community
langchain-text-splitters
langchain

# Vector store
chromadb

# BM25 for hybrid search
rank_bm25

# Cross-encoder for reranking and evaluation
sentence-transformers

# Environment variables
python-dotenv

# PDF processing (optional)
PyPDF2

# Frontend
streamlit

# Additional dependencies
tiktoken
